Completed refactoring the details form the plots. All plots are identical save selector classes. Refactoring that out next.
