+++ client/src/components/ProTx/components/charts/AnalyticsDetails.css
@@ -1,0 +1,1 @@
+/* Analytics Details */
