--- client/src/components/ProTx/components/charts/PredictiveFeaturesChart.css
@@ -1,1 +1,0 @@
-/* .predictive-features-chart {} */
